admin e implementação do dashboard

+++ app/Http/Controllers/ContatosController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Contato;
 use Session;
@@ -38,7 +39,13 @@ class ContatosController extends Controller
      */
     public function create()
     {
-       return view('contato.create');
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            return view('contato.create');
+        }
+        else {
+            return redirect('login');
+        }
+
     }
 
     /**
@@ -49,24 +56,33 @@ class ContatosController extends Controller
      */
     public function store(Request $request)
     {
-        $contato = new Contato();
-        $contato->nome = $request->input('nome');
-        $contato->email = $request->input('email');
-        $contato->telefone = $request->input('telefone');
-        $contato->cidade = $request->input('cidade');
-        $contato->estado = $request->input('estado');
-        if($contato->save()){
-            if($request->hasFile('foto')){
-                $imagem = $request->file('foto');
-                $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
-                //dd($imagem, $nomearquivo,$contato->id);
-                $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
-            }
-            return redirect('contatos');
-            
-        }
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $this->validate($request,[
+                'nome' => 'required|min:3',
+                'email' => 'required|e-mail',
+                'telefone' => 'required',
+                'cidade' => 'required',
+                'estado' => 'required',
+            ]);
+            $contato = new Contato();
+            $contato->nome = $request->input('nome');
+            $contato->email = $request->input('email');
+            $contato->telefone = $request->input('telefone');
+            $contato->cidade = $request->input('cidade');
+            $contato->estado = $request->input('estado');
+            if($contato->save()) {
+                if($request->hasFile('foto')){
+                    $imagem = $request->file('foto');
+                    $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
+                    //dd($imagem, $nomearquivo,$contato->id);
+                    $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
+                }
+                return redirect('contatos');
+            } else {
+                return redirect('login');
     }
-
+    }
+    }
     /**
      * Display the specified resource.
      *
@@ -87,8 +103,12 @@ class ContatosController extends Controller
      */
     public function edit($id)
     {
-        $contato = Contato::find($id);
-        return view('contato.edit', array('contato'=> $contato));
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $contato = Contato::find($id);
+            return view('contato.edit',array('contato' => $contato));
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -100,21 +120,32 @@ class ContatosController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $contato = Contato::find($id);
-        if($request->hasFile('foto')){
-            $imagem = $request->file('foto');
-            $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
-            $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $this->validate($request,[
+                'nome' => 'required|min:3',
+                'email' => 'required|e-mail|min:3',
+                'telefone' => 'required',
+                'cidade' => 'required',
+                'estado' => 'required',
+            ]);
+            $contato = Contato::find($id);
+            if($request->hasFile('foto')){
+                $imagem = $request->file('foto');
+                $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
+                $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
+            }
+            $contato->nome = $request->input('nome');
+            $contato->email = $request->input('email');
+            $contato->telefone = $request->input('telefone');
+            $contato->cidade = $request->input('cidade');
+            $contato->estado = $request->input('estado');
+            if($contato->save()) {
+                Session::flash('mensagem','Contato alterado com sucesso');
+                return redirect()->back();
+            }
+        } else {
+            return redurect('login');
         }
-        $contato->nome = $request->input('nome');
-        $contato->email = $request->input('email');
-        $contato->telefone = $request->input('telefone');
-        $contato->cidade = $request->input('cidade');
-        $contato->estado = $request->input('estado');
-        if($contato->save()){
-            Session::flash('mensagem', 'Contato alterado com sucesso');
-            return redirect(url('contatos/'));
-    }
     }
     /**
      * Remove the specified resource from storage.
@@ -124,9 +155,16 @@ class ContatosController extends Controller
      */
     public function destroy($id)
     {
-        $contato = Contato::find($id);
-        $contato->delete();
-        Session::flash('mensagem','Contato excluído com sucesso');
-        return redirect(url('contatos/'));
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $contato = Contato::find($id);
+            if (isset($request->foto)) {
+            unlink($request->foto);
+            }
+            $contato->delete();
+            Session::flash('mensagem','Contato Excluído com Sucesso Foto:');
+            return redirect(url('contatos/'));
+        } else {
+            return redirect('login');
     }
+}
 }
