+++ app/Http/Controllers/EmprestimosController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Models\Emprestimo;
 use App\Models\Livro;
 use App\Models\Contato;
@@ -19,6 +20,7 @@ class EmprestimosController extends Controller
     {
         $emprestimos = Emprestimo::paginate(5);
         return view('emprestimo.index',array('emprestimos' => $emprestimos,'busca'=>null));
+        return redirect('login');
     }
 
     /**
@@ -26,19 +28,35 @@ class EmprestimosController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-        public function buscar(Request $request) {
-        $emprestimos = Emprestimo::with('contato')->with('livro')->where('contato_id','=',$request->input('busca'))
-        ->orwhere('livro_id','=',$request->input('busca'))->orwhere('obs','LIKE','%'.$request->input('busca').'%')
-        ->paginate(5); return view('emprestimo.index',array('emprestimos' => $emprestimos,'busca'=>$request->input('busca')));
-    }
-
 
     public function create()
     {
-        $contatos = Contato::all();
-        $livros = Livro::all();
-       return view('emprestimo.create',['contatos'=>$contatos,
-       'livros'=>$livros]);
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $contatos = Contato::all();
+            $livros = Livro::all();
+            return view('emprestimo.create',['contatos'=>$contatos,'livros'=>$livros]);
+        } else {
+            return redirect('login');
+        }
+    }
+    
+ /**
+     * Store a newly created resource in storage.
+     *
+     *
+     * @return \Illuminate\Http\Response
+     */
+
+    public function buscar(Request $request) {
+        $emprestimos = Emprestimo::join('contatos','contatos.id','=','emprestimos.contato_id')
+                    ->join('livros','livros.id','=','emprestimos.livro_id')
+                    ->select('emprestimos.*','contatos.nome','livros.titulo')
+                    ->where('contato_id','=',$request->input('busca'))
+                    ->orwhere('livro_id','=',$request->input('busca'))
+                    ->orwhere('obs','LIKE','%'.$request->input('busca').'%')->orwhere('contatos.nome','LIKE','%'.$request->input('busca').'%')
+                    ->orwhere('livros.titulo','LIKE','%'.$request->input('busca').'%')
+                    ->simplepaginate(5);
+        return view('emprestimo.index',array('emprestimos' => $emprestimos,'busca'=>$request->input('busca')));
     }
 
     /**
@@ -49,20 +67,26 @@ class EmprestimosController extends Controller
      */
     public function store(Request $request)
     {
-        $emprestimo = new Emprestimo();
-        $emprestimo->contato_id = $request->input('contato_id');
-        $emprestimo->livro_id = $request->input('livro_id');
-        $emprestimo->datahora=
-        \Carbon\Carbon::createFromFormat('d/m/Y H:i:s',
-        $request->input('datahora'));
-        $emprestimo->obs = $request->input('obs');
-        $emprestimo->datadevolucao = null;
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $this->validate($request,[
+                'contato_id' => 'required',
+                'livro_id' => 'required',
+                'datahora' => 'required'
+            ]);
+            $emprestimo = new Emprestimo();
+            $emprestimo->contato_id = $request->input('contato_id');
+            $emprestimo->livro_id = $request->input('livro_id');
+            $emprestimo->datahora = \Carbon\Carbon::createFromFormat('d/m/Y H:i:s', $request->input('datahora'));
+            $emprestimo->obs = $request->input('obs');
+            $emprestimo->datadevolucao = null;
 
-        if($emprestimo->save()){
-            return redirect('emprestimos');
-        }
+            if($emprestimo->save()) {
+                return redirect('emprestimos');
+            }
+        } else {
+            return redirect('login');
     }
-
+    }
     /**
      * Display the specified resource.
      *
@@ -97,16 +121,19 @@ class EmprestimosController extends Controller
 
     public function devolver(Request $request, $id)
     {
-        $emprestimo = Emprestimo::find($id);
-        $emprestimo->datadevolucao = \Carbon\Carbon::now();
-        $emprestimo->save();
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $emprestimo = Emprestimo::find($id);
+            $emprestimo->datadevolucao = \Carbon\Carbon::now();
+            $emprestimo->save();
 
-        if($emprestimo->save()) {
-            Session::flash('mensagem','Empréstimo Devolvido');
-            return redirect('/emprestimos');
-        }
+            if($emprestimo->save()) {
+                Session::flash('mensagem','Empréstimo Devolvido');
+                return redirect()->back();
+            }
+        } else {
+            return redirect('login');
     }
-
+}
     /**
      * Update the specified resource in storage.
      *
@@ -127,10 +154,16 @@ class EmprestimosController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        $emprestimo = Emprestimo::find($id);
+        if ((Auth::check()) && (Auth::user()->isAdmin())) {
+            $emprestimo = Emprestimo::find($id);
 
-        $emprestimo->delete();
-        Session::flash('mensagem','Empréstimo excluído com sucesso');
-        return redirect(url('emprestimos/'));
+            $emprestimo->delete();
+            Session::flash('mensagem','Empréstimo Excluído com Sucesso');
+            return redirect(url('emprestimos/'));
+        } else {
+            return redirect('login');
+        }
     }
-}
+    }
+
+
