+++ resources/views/emprestimo/show.blade.php
@@ -1,5 +1,5 @@
 
-@extends('layout.app')
+@extends('layouts.app')
 @section('title','Empréstimo - '.$emprestimo->id)
 @section('content')
 <br><br>
@@ -20,9 +20,11 @@
                     </div>
                     <div class="col-4">
                         @if($emprestimo->datadevolucao==null)
+                        @if ((Auth::check()) && (Auth::user()->isAdmin()))
                         {{Form::open(['route'=>['emprestimos.devolver',$emprestimo->id],'method'=>'PUT'])}}
                         {{form::submit('Devolver',['class'=>'btn btn-success','onclick'=>'return confim("Confirma devolução?")'])}}
                         {{Form::close()}}
+                        @endauth
                         @endif
                     </div>
                 </div>
@@ -35,11 +37,14 @@
             <p class="text">obs: {{$emprestimo->obs}}</p>
         </div>
         <div class="card-footer">
+        @if ((Auth::check()) && (Auth::user()->isAdmin()))
             {{Form::open(['route' => ['emprestimos.destroy',$emprestimo->id],'method' => 'DELETE'])}}
             {{Form::submit('Excluir',['class'=>'btn btn-danger','onclick'=>'return confirm("Confirma exclusão?")'])}}
+            @endauth
             <a href="{{url('emprestimos/')}}" class="btn btn-warning">Voltar</a>
+            @if ((Auth::check()) && (Auth::user()->isAdmin()))
             {{Form::close()}}
-          
+            @endauth
         </div>
     </div>
 @endsection
